Unseparated activitypub views and models
