+++ democrasite/activitypub/views.py
@@ -2,6 +2,7 @@ from functools import wraps
 from http import HTTPStatus
 from typing import TYPE_CHECKING
 
+from django import forms
 from django import http
 from django.contrib import messages
 from django.contrib.auth.mixins import UserPassesTestMixin
@@ -19,8 +20,6 @@ from django.views.generic import DetailView
 from django.views.generic import ListView
 from django.views.generic import UpdateView
 
-from .forms import NoteForm
-from .forms import PersonForm
 from .models import Note
 from .models import Person
 
@@ -76,6 +75,22 @@ class NoteDetailView(DetailView):
 
 
 note_detail_view = NoteDetailView.as_view()
+
+
+class NoteForm(forms.ModelForm):
+    """Form for creating or replying to a note."""
+
+    content = forms.CharField(
+        max_length=500,
+        widget=forms.Textarea(
+            attrs={"rows": 3, "placeholder": "Share your thoughts..."}
+        ),
+        help_text="Your note content (max 500 characters).",
+    )
+
+    class Meta:
+        model = Note
+        fields = ["content"]
 
 
 class NoteCreateView(UserProfileMixin, CreateView):
@@ -179,6 +194,21 @@ class PersonCreateView(SuccessMessageMixin, CreateView):
 
 
 person_create_view = PersonCreateView.as_view()
+
+
+class PersonForm(forms.ModelForm):
+    """Form for creating or updating a person's profile."""
+
+    bio = forms.CharField(
+        max_length=500,
+        required=False,
+        widget=forms.Textarea(attrs={"rows": 3, "placeholder": "Describe yourself"}),
+        help_text="Your bio content (max 500 characters).",
+    )
+
+    class Meta:
+        model = Person
+        fields = ["bio"]
 
 
 class PersonUpdateView(UserProfileMixin, UpdateView):
